pyproject.toml
[project]
name = "event-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "azure-core==1.38.2",
    "azure-eventhub==5.15.1",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "dotenv==0.9.9",
    "faker==40.5.1",
    "fastapi[standard]==0.133.1",
    "h11==0.16.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "markupsafe==3.0.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "starlette==0.52.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uvicorn==0.41.0",
]
